Add user header
Expose a clickable user name in the top bar and wire user state across screens: pull current user from UserManager and pass onUserClick handlers into Battery, Home, Performance, Social, Settings, Documentation and Profile screens.

+++ app/src/main/java/com/example/myfulgora/data/remote/FulgoraMqttClient.kt
@@ -75,7 +75,7 @@ object FulgoraMqttClient {
                     isOnline = json.optBoolean("isOnline", currentState.isOnline),
                     isCharging = json.optBoolean("isCharging", currentState.isCharging),
                     batteryCycles = json.optInt("batteryCycles", currentState.batteryCycles),
-                    batteryTemp = tempDouble.toInt(),
+                    batteryTemp = tempDouble.toInt().toDouble(),
                     batteryHealth = json.optString("batteryHealth", currentState.batteryHealth),
                     timeLeft = json.optString("timeLeft", currentState.timeLeft)
                 )
